Gestion membre

+++ admin/gestion_membres.php
@@ -11,15 +11,16 @@ if($_POST){
 	
 	//debug($_POST);
 
-	if(isset($_POST['mdp']) && !empty($_POST['mdp'])){		
+	if(isset($_POST['mdp']) && !empty($_POST['mdp'])){
+			$contenu .= '<h2> Ajout et modification d\'un memebre</h2>';		
 		
-		if(isset($_GET['action']) && $_GET['action'] == 'modifier'){
-			$resultat = $pdo -> prepare("REPLACE INTO membre (id_membre, pseudo, mdp, nom, prenom, email, civilite, ville, code_postal, adresse, statut) VALUES (:id_membre, :pseudo, :mdp, :nom, :prenom, :email, :civilite, :ville, :code_postal, :adresse, :statut)");
+		if(isset($_GET['action']) && $_GET['action'] == 'modifier'){//modif
+			$resultat = $pdo -> prepare("REPLACE INTO membre (id_membre, pseudo, mdp, nom, prenom, email, civilite, statut, date_enregistrement) VALUES (:id_membre, :pseudo, :mdp, :nom, :prenom, :email, :civilite, :statut, NOW())");
 			
 			$resultat -> bindParam(':id_membre', $_POST['id_membre'], PDO::PARAM_INT);
 		}
-		else{
-			$resultat = $pdo -> prepare("REPLACE INTO membre (pseudo, mdp, nom, prenom, email, civilite, ville, code_postal, adresse, statut) VALUES (:pseudo, :mdp, :nom, :prenom, :email, :civilite, :ville, :code_postal, :adresse, :statut)");
+		else{//ajout
+			$resultat = $pdo -> prepare("REPLACE INTO membre (pseudo, mdp, nom, prenom, email, civilite, statut, date_enregistrement) VALUES (:pseudo, :mdp, :nom, :prenom, :email, :civilite, :statut, NOW())");
 		}
 		
 		//STR
@@ -30,12 +31,7 @@ if($_POST){
 		$resultat -> bindParam(':prenom', $_POST['prenom'], PDO::PARAM_STR);
 		$resultat -> bindParam(':email', $_POST['email'], PDO::PARAM_STR);
 		$resultat -> bindParam(':civilite', $_POST['civilite'], PDO::PARAM_STR);
-		$resultat -> bindParam(':ville', $_POST['ville'], PDO::PARAM_STR);
 		$resultat -> bindParam(':statut', $_POST['statut'], PDO::PARAM_STR);
-		$resultat -> bindParam(':adresse', $_POST['adresse'], PDO::PARAM_STR);
-		
-		//INT
-		$resultat -> bindParam(':code_postal', $_POST['code_postal'], PDO::PARAM_INT);
 		
 		if($resultat -> execute()){
 			$_GET['action'] = 'affichage';
@@ -60,8 +56,10 @@ if(isset($_GET['action']) && $_GET['action'] == 'supprimer'){ // SI une
 		
 		if($resultat -> rowCount() > 0){
  
-			$resultat = $pdo -> exec("DELETE FROM membre WHERE id_membre = $membre[id_membre]");
-			
+			$resultat = $pdo -> prepare(" DELETE FROM membre WHERE id_membre = :id_membre ");
+			$resultat -> bindParam(':id_membre', $_GET['id_membre'], PDO::PARAM_INT);
+			$resultat -> execute();
+
 			if($resultat != FALSE){
 				$_GET['action'] = 'affichage';
 				$msg .= '<div class="validation">Le membre N°' . $membre['id_membre'] . ' a bien été supprimé !</div>';	
@@ -75,7 +73,7 @@ if(isset($_GET['action']) && $_GET['action'] == 'affichage'){
 
 	$resultat = $pdo -> query("SELECT * FROM membre"); 
 	
-
+	$contenu .= '<h2>Liste des membres</h2>';
 	$contenu .= '<table border="1">';
 	$contenu .= '<tr>';
 	for($i = 0; $i < $resultat -> columnCount(); $i++){
@@ -94,7 +92,7 @@ if(isset($_GET['action']) && $_GET['action'] == 'affichage'){
 			}
 		} 
 		$contenu .= '<td><a href="?action=modifier&id_membre='. $membres['id_membre'] .'"><img src="' . RACINE_SITE . 'img/edit.png"/></a></td>';
-		$contenu .= '<td><a href="?action=supprimer&id_membre='. $membres['id_membre'] .'"><img src="' . RACINE_SITE . 'img/delete.png"/></a></td>';
+		$contenu .= '<td><a href="?action=supprimer&id_membre='. $membres['id_membre'] .'" onclick="return confirm(\'Voulez-vous supprimer le membre ' . $membres['pseudo'] . ' ?\');"><img src="' . RACINE_SITE . 'img/delete.png"/></a></td>';
 		$contenu .= '</tr>';
 	}
 	$contenu .= '</table>';
@@ -128,9 +126,6 @@ $prenom = (isset($membre_actuel)) ? $membre_actuel['prenom'] : '';
 $email = (isset($membre_actuel)) ? $membre_actuel['email'] : '';
 $civilite = (isset($membre_actuel)) ? $membre_actuel['civilite'] : '';
 $statut = (isset($membre_actuel)) ? $membre_actuel['statut'] : '';
-$ville = (isset($membre_actuel)) ? $membre_actuel['ville'] : '';
-$adresse = (isset($membre_actuel)) ? $membre_actuel['adresse'] : '';
-$code_postal = (isset($membre_actuel)) ? $membre_actuel['code_postal'] : '';
 $action = (isset($membre_actuel)) ? 'Modifier' : 'Ajouter';
 $id_membre = (isset($membre_actuel)) ? $membre_actuel['id_membre'] : '';
 ?>
@@ -139,44 +134,34 @@ $id_membre = (isset($membre_actuel)) ? $membre_actuel['id_membre'] : '';
 <h2><?= $action ?> un membre</h2>
 
 <form method="post" action="" enctype="multipart/form-data">
-	<input type="hidden" name="id_membre" value="<?= $id_membre ?>" />
+	<input type="hidden" name="id_membre" value="<?= $id_membre ?>" required />
 
 	
 	<label>Pseudo : </label>
-	<input type="text" name="pseudo" value="<?= $pseudo ?>" /><br/>
+	<input type="text" name="pseudo" value="<?= $pseudo ?>"  required/><br/>
 	
 	<label>Mot de passe: </label>
-	<input type="text" name="mdp"/><br/>
+	<input type="text" name="mdp" required/><br/>
 	
 	<label>Nom: </label>
-	<input type="text" name="nom" value="<?= $nom ?>"/><br/>
+	<input type="text" name="nom" value="<?= $nom ?>" required/><br/>
 	
 	<label>Prénom: </label>
-	<textarea name="prenom"><?= $prenom ?></textarea><br/>
+	<input name="prenom" value="<?= $prenom ?>" required/><br/>
 	
 	<label>Email: </label>
-	<input type="text" name="email" value="<?= $email ?>"/><br/>
+	<input type="email" name="email" value="<?= $email ?>" required/><br/>
 	
 	<label>Civilite: </label>
-	<select name="civilite">
-		<option>-- Selectionnez --</option>
+	<select name="civilite" required>
+		<option value="">-- Selectionnez --</option>
 		<option <?= ($civilite == 'm') ? 'selected' : '' ?> value="m">Homme</option>
 		<option <?= ($civilite == 'f') ? 'selected' : '' ?> value="f">Femme</option>
 	</select><br/>
 	
-
-	<label>Ville: </label>
-	<input type="text" name="ville" value="<?= $ville ?>"/><br/>
-	
-	<label>Code Postal : </label>
-	<input type="text" name="code_postal" value="<?= $code_postal ?>"/><br/>
-	
-	<label>Adresse : </label>
-	<input type="text" name="adresse" value="<?= $adresse ?>"/><br/>
-	
 	<label>Statut: </label>
-	<select name="statut">
-		<option>-- Selectionnez --</option>
+	<select name="statut" required>
+		<option value="">-- Selectionnez --</option>
 		<option <?= ($statut == '0') ? 'selected' : '' ?> value="0">Membre</option>
 		<option <?= ($civilite == '1') ? 'selected' : '' ?> value="1">Admin</option>
 	</select><br/>
